CryptBasicDecompressor#flushBuffer(byte[] buf, int off, int len) was computing wrong the size of buffer to copy.
buf is the destination buffer to copy into.
off is the position in buf to copy into.
len is the desired number of bytes.
size is a number <= len, telling how much bytes from "out" can be copied.

The number of bytes that cam be copied is:

    Math.min(Math.min(len, buf.length - off), out.position());

the minimum of:
- the desired length (len)
the number of elements that can go inside the buffer (buf.length-off)
- OR the number of elements that are inside of the inner buffer to copy from (out).

you copy, and you return the actual number of bytes copied.

The former equation was:

    Math.min(Math.min(len, buf.length) - off, out.position());

what for example makes no sense in the case of wanting to copy 1 last byte into offset 4 => wrong size=-3.